# Implement CanPlay and Moveable