--- a/core/src/main/scala/HasPosition.scala
+++ b/core/src/main/scala/HasPosition.scala
@@ -1,0 +1,10 @@
+package chess
+
+import chess.variant.Variant
+
+trait HasPosition[A]:
+  extension (a: A)
+    def position: Position
+    inline def variant: Variant = position.variant
+    inline def color: Color     = position.color
+    inline def history: History = position.history