- `Moveable` is a super trait of `Uci` and `San`, which can be applied to a Position and may return a new `MoveOrDrop`.
- `CanPlay` is a type that can apply a Moveable and may return a `MoveOrDrop` and its next state. Currently, `Position`, `Game` and `Position.AndFullMoveNumber` are implemented with `CanPlay`.

These additions leads to more consistent and flexible api without Replay which in turn leads to better performance.

The best example of those improvement is: https://github.com/lichess-org/lila/pull/17551/files#diff-543266d16980131108154ae76e3b334d7036f0e374d1cea7ad8c006d92e3f8c0

master:
```
[info] Benchmark           Mode  Cnt     Score    Error  Units
[info] PlayBench.divider  thrpt   45   163.634 ±  2.920  ops/s
[info] PlayBench.play     thrpt   45  3038.565 ± 15.615  ops/s
[info] PlayBench.replay   thrpt   45   134.144 ±  0.542  ops/s
```
pr:
```
[info] Benchmark           Mode  Cnt     Score    Error  Units
[info] PlayBench.divider  thrpt   45   173.623 ±  0.299  ops/s
[info] PlayBench.play     thrpt   45  3061.146 ± 13.219  ops/s
[info] PlayBench.replay   thrpt   45   138.284 ±  0.787  ops/s
```